notifications refactoring
