+++ types/context/notificationcontext.ts
@@ -1,0 +1,7 @@
+import { notifications } from "@mantine/notifications";
+
+/** 
+ * Type declaration for the notification context. 
+ * This could be abstracted from mantine if necessary. 
+ */
+export type NotificationContextType = typeof notifications;
